# openapi.yaml
openapi: 3.0.0
info:
  title: WeatherOps API
  description: >
    Event-driven weather intelligence platform for real-time monitoring,
    alert management, and automated weather-triggered actions.

    ## Features
    - User authentication with JWT tokens
    - Weather location monitoring
    - Custom weather alert rules
    - Real-time alert management
    - Multi-channel notifications
    - User preferences and settings

    ## Base URL
    Production: `https://api.weatherops.local/api/v1`
    Development: `http://localhost:8000/api/v1`
  version: 1.1.0
  contact:
    name: WeatherOps Support
    email: support@weatherops.local
  license:
    name: Apache 2.0
  x-logo:
    url: https://weatherops.local/logo.png
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://api.weatherops.local/api/v1
    description: Production server
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
  schemas:
...
    RuleCreate:
      type: object
      required:
        - location_id
        - metric
        - operator
        - threshold
      properties:
        location_id:
          type: string
          format: uuid
          description: ID of the location this rule monitors
          example: 550e8400-e29b-41d4-a716-446655440000
        metric:
          type: string
          enum:
            - temperature
            - rainfall
            - wind_speed
            - humidity
          description: Weather metric to monitor
          example: temperature
        operator:
          type: string
          enum:
            - ">"
            - "<"
            - ">="
            - "<="
            - "=="
          description: Comparison operator for threshold
          example: ">"
        threshold:
          type: number
          format: double
          description: Threshold value that triggers the alert
          example: 35.0
      description: Create new weather rule request
    RuleUpdate:
      type: object
      properties:
        metric:
          type: string
          enum:
            - temperature
            - rainfall
            - wind_speed
            - humidity
          description: Weather metric to monitor
          example: temperature
        operator:
          type: string
          enum:
            - ">"
            - "<"
            - ">="
            - "<="
            - "=="
          description: Comparison operator for threshold
          example: ">="
        threshold:
          type: number
          format: double
          description: Threshold value that triggers the alert
          example: 38.0
        is_active:
          type: boolean
          description: Enable or disable this rule
          example: true
      description: Update weather rule request
    RuleResponse:
      type: object
      required:
        - id
        - location_id
        - metric
        - operator
        - threshold
        - is_active
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Rule unique identifier
          example: 550e8400-e29b-41d4-a716-446655440000
        location_id:
          type: string
          format: uuid
          description: Location ID
          example: 660e8400-e29b-41d4-a716-446655440000
        metric:
          type: string
          enum:
            - temperature
            - rainfall
            - wind_speed
            - humidity
          description: Weather metric being monitored
          example: temperature
        operator:
          type: string
          enum:
            - ">"
            - "<"
            - ">="
            - "<="
            - "=="
          description: Comparison operator
          example: ">"
        threshold:
          type: number
          format: double
          description: Alert threshold value
          example: 35.0
        is_active:
          type: boolean
          description: Whether this rule is currently active
          example: true
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
          example: 2024-06-05T10:30:00Z
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
          example: 2024-06-05T10:30:00Z
      description: Weather rule information response

    # Alert Schemas
...
    UserPreferenceResponse:
      type: object
      required:
        - user_id
        - notification_channels
        - email_notifications_enabled
        - sms_notifications_enabled
        - webhook_notifications_enabled
      properties:
        user_id:
          type: string
          format: uuid
          description: User ID
          example: 550e8400-e29b-41d4-a716-446655440000
        notification_channels:
          type: array
          items:
            type: string
            enum:
              - email
              - sms
              - webhook
              - system
          description: Enabled notification channels
          example:
            - email
            - webhook
        email_notifications_enabled:
          type: boolean
          description: Email notifications enabled
          example: true
        sms_notifications_enabled:
          type: boolean
          description: SMS notifications enabled
          example: false
        webhook_notifications_enabled:
          type: boolean
          description: Webhook notifications enabled
          example: true
        webhook_url:
          type: string
          format: uri
          nullable: true
          description: Webhook endpoint URL
          example: https://example.com/webhooks/alerts
        alert_threshold:
          type: number
          format: double
          nullable: true
          description: Custom alert threshold
          example: 35.0
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
          example: 2024-06-05T10:30:00Z
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
          example: 2024-06-05T10:30:00Z
      description: User preference settings response
    UserPreferenceUpdate:
      type: object
      properties:
        notification_channels:
          type: array
          items:
            type: string
            enum:
              - email
              - sms
              - webhook
              - system
          description: Enabled notification channels
          example:
            - email
            - webhook
        email_notifications_enabled:
          type: boolean
          description: Enable email notifications
          example: true
        sms_notifications_enabled:
          type: boolean
          description: Enable SMS notifications
          example: false
        webhook_notifications_enabled:
          type: boolean
          description: Enable webhook notifications
          example: true
        webhook_url:
          type: string
          format: uri
          nullable: true
          description: Webhook endpoint URL
          example: https://example.com/webhooks/alerts
        alert_threshold:
          type: number
          format: double
          nullable: true
          description: Custom alert threshold
          example: 35.0
      description: Update user preferences request

    # Error Responses
    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message
          example: Invalid credentials
      description: Error response
    ValidationErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  oneOf:
                    - type: string
                    - type: integer
              msg:
                type: string
              type:
                type: string
          description: Validation error details
      description: Validation error response
...
paths:
...
  /rules:
    post:
      tags:
        - Rules
      summary: Create Rule
      description: Create a new weather alert rule for a location
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleCreate'
      responses:
        '201':
          description: Rule created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleResponse'
        '400':
          description: Invalid rule data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - location access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
      operationId: create_rule
  /rules/location/{location_id}:
    get:
      tags:
        - Rules
      summary: Get Location Rules
      description: Get all alert rules for a specific location
      parameters:
        - name: location_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Location ID
      responses:
        '200':
          description: Rules retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RuleResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Location not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: get_location_rules
  /rules/{rule_id}:
    get:
      tags:
        - Rules
      summary: Get Rule
      description: Get details for a specific rule
      parameters:
        - name: rule_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Rule ID
      responses:
        '200':
          description: Rule retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Rule not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: get_rule

    patch:
      tags:
        - Rules
      summary: Update Rule
      description: Update rule configuration (metric, operator, threshold, or active status)
      parameters:
        - name: rule_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Rule ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleUpdate'
      responses:
        '200':
          description: Rule updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Rule not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
      operationId: update_rule

    delete:
      tags:
        - Rules
      summary: Delete Rule
      description: Delete a weather alert rule
      parameters:
        - name: rule_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Rule ID
      responses:
        '204':
          description: Rule deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Rule not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: delete_rule
...
  /preferences:
    get:
      tags:
        - Preferences
      summary: Get User Preferences
      description: Get user notification preferences and settings
      responses:
        '200':
          description: User preferences retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferenceResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Preferences not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: get_preferences

    patch:
      tags:
        - Preferences
      summary: Update User Preferences
      description: Update notification channels, webhooks, and thresholds
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPreferenceUpdate'
      responses:
        '200':
          description: User preferences updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferenceResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
      operationId: update_preferences

    delete:
      tags:
        - Preferences
      summary: Delete User Preferences
      description: Reset user preferences to defaults
      responses:
        '204':
          description: User preferences deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: delete_preferences